Fix ill-formed return value in PySocket, threading issues in benchmark.py
PySocket::recv would return an ill-formed Python object if recv() returned
with size==-1. A tuple with size==-1 and None fixes the problem.
In addition, benchmark.py would hang if run with `-c thread` due to Python's
global interpreter lock. An alternative implementation with `try_recv_raw`
was added to fix the issue.